Slanter

Slanter — Module

Reorder matrices rows and columns to move high values close to the diagonal.

Slanter.slanted_orders — Function

function slanted_orders(
    data::AbstractMatrix{<:Real};
    order_rows::Bool=true,
    order_cols::Bool=true,
    squared_order::Bool=true,
    same_order::Bool=false,
    discount_outliers::Bool=true,
    max_spin_count::Integer=10
)::Tuple{AbstractVector{<:Integer}, AbstractVector{<:Integer}}

Compute rows and columns orders which move high values close to the diagonal.

For a matrix expressing the cross-similarity between two (possibly different) sets of entities, this produces better results than clustering. This is because clustering does not care about the order of each two sub-partitions. That is, clustering is as happy with ((2, 1), (4, 3)) as it is with the more sensible ((1, 2), (3, 4)). As a result, visualizations of similarities using naive clustering can be misleading.

This situation is worse in Python or R than it is in Julia, which mercifully provides the :barjoseph method to reorder the branches. Still, the method here may be "more suitable" in specific circumstances (when the data depicts a clear gradient).

Parameters

data: A rectangular matrix containing non-negative values (may be negative if squared_order).

order_rows: Whether to reorder the rows.

order_cols: Whether to reorder the columns.

squared_order: Whether to reorder to minimize the l2 norm (otherwise minimizes the l1 norm).

same_order: Whether to apply the same order to both rows and columns.

discount_outliers: Whether to do a final order phase discounting outlier values far from the diagonal.

max_spin_count: How many times to retry improving the solution before giving up.

Returns

A tuple with two vectors, which contain the order of the rows and the columns.

Slanter.slanted_reorder — Function

slanted_reorder(
    data::AbstractMatrix{T};
    order_data::Union{AbstractMatrix{<:Real}, Nothing}=nothing,
    order_rows::Bool=true,
    order_cols::Bool=true,
    squared_order::Bool=true,
    same_order::Bool=false,
    discount_outliers::Bool=true
)::AbstractMatrix{T} where {T<:Real}

Reorder data rows and columns to move high values close to the diagonal.

Given a matrix expressing the cross-similarity between two (possibly different) sets of entities, this uses slanted_orders to compute the "best" order for visualizing the matrix, then returns the reordered data.

Parameters

data: A rectangular matrix to reorder, of non-negative values (unless order_data is specified, or squared_order)).

order_data: An optional matrix of non-negative values of the same size to use for computing the orders (may be negative if squared_order).

order_rows: Whether to reorder the rows.

order_cols: Whether to reorder the columns.

squared_order: Whether to reorder to minimize the l2 norm (otherwise minimizes the l1 norm).

same_order: Whether to apply the same order to both rows and columns.

discount_outliers: Whether to do a final order phase discounting outlier values far from the diagonal.

Returns

A matrix of the same shape whose rows and columns are a permutation of the input.

Slanter.reorder_hclust — Function

reorder_hclust(clusters, order)

Given a clustering of some data, and some ideal order we'd like to use to visualize it, reorder (but do not modify) the clustering to be as consistent as possible with this ideal order.

Parameters

clusters: The existing clustering of the data.

order: The ideal order we'd like to see the data in.

Returns

A reordered clustering which is consistent, wherever possible, with the ideal order.

Slanter.oclust — Function

oclust(distances; method="ward.D2", order=nothing, members=nothing)

Hierarchically cluster ordered data.

Given a distance matrix for sorted objects, compute a hierarchical clustering preserving this order. That is, this is similar to hclust with the constraint that the result's order is always 1:N.

If an order is specified, assumes that the data will be re-ordered by this order. That is, the indices in the returned clustering object will refer to the post-reorder data locations, not to the current data locations.

Currently, the only methods supported are ward.D and ward.D2.

Parameters

distances: A distance matrix.

method: The clustering method to use (only ward.D and ward.D2 are supported).

order: If specified, assume the data will be re-ordered by this order.

members: Optionally, the number of members for each row/column of the distances (by default, one each).

Returns

A clustering object (similar to R's hclust).

Index

Slanter

Slanter.oclust

Slanter.reorder_hclust

Slanter.slanted_orders

Slanter.slanted_reorder